backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
mysqlclient==2.2.4
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1